% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Render a Git repository report}
\usage{
report(
  after,
  until = Sys.Date(),
  max = 1e+09,
  repo = ".",
  output_dir = tempdir(),
  output_file = "report.html",
  files_stats = FALSE,
  render_quiet = TRUE
)
}
\arguments{
\item{after}{character. Starting date.}

\item{until}{character. Ending data. Defaults to today.}

\item{max}{numeric. Maximum number of commits to retrieve. Defaults to 1e+9.}

\item{repo}{character. Repository path on your computer. Defaults to the active repository. You can also provide an SSH repository address that will be temporally cloned.}

\item{output_dir}{character. Output directory. Defaults to temporary dir.}

\item{output_file}{character. Output report name. Defaults to \code{report.html}.}

\item{files_stats}{logical. Fetch file stats. Can take longer on large repos. Defaults to \code{FALSE}.}

\item{render_quiet}{logical. Show report rendering messages. Defaults to \code{FALSE}.}
}
\value{
an html report file.
}
\description{
Render a Git repository report
}
